refactor: remove as, move tests to component folder, props
