+++ src/components/cardList/CardList.tsx
@@ -4,11 +4,9 @@ import Card from 'components/card/Card';
 
 import styles from './CardList.module.scss';
 
-export default function CardList({
-  data = [],
-}: {
-  data: Product[] | undefined;
-}) {
+type CardListProps = { data?: Product[] };
+
+export default function CardList({ data = [] }: CardListProps) {
   if (!data.length) {
     return (
       <div className={styles.container}>
